clean up assignment 2
A mother kangaroo has lost some number of her baby kangroos ("joeys") in a shopping mall. To be easy to find she will stay outside. Each joey was lost in its own unique location, and thus must make its own way out of the mall.

To find its way out of the mall, a joey must hop along the corridor it is in and then choose the proper direction when the corridor ends (either left or right) by sending either LEFT_TURN_SIGNAL (SIGUSR1) or RIGHT_TURN_SIGNAL (SIGUSR2) to the mall. If it chooses incorrectly then the mall will instantly tell it whether or not it has made the correct turn.